requirements.txt
asgiref==3.7.2
Django==3.2.23
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
mysqlclient==2.2.0
python-http-client==3.3.7
pytz==2023.3.post1
sendgrid==6.11.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
Django==3.2.18
django-environ==0.11.2
exceptiongroup==1.1.3
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
install==1.3.5
multidict==6.0.4
mysql-connector-python==8.0.33
mysqlclient==2.2.4
openai==0.27.4
outcome==1.2.0
protobuf==3.20.3
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
PySocks==1.7.1
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2022.7.1
regex==2023.5.5
requests==2.29.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.10.0
sqlparse==0.4.3
tiktoken==0.4.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.4
urllib3==1.26.15
wsproto==1.2.0
yarl==1.9.2
